add resize function fix expend function crash when down scale the size add rule settings

+++ srcs/gol_get_at.c
@@ -6,7 +6,7 @@
 /*   By: gcros <gcros@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/22 21:40:16 by gcros             #+#    #+#             */
-/*   Updated: 2025/03/18 16:54:46 by gcros            ###   ########.fr       */
+/*   Updated: 2025/08/17 22:00:42 by gcros            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@ int	gol_get_at(t_gol *gol, size_t x, size_t y)
 	return (gol->cells[y * gol->width + x].value);
 }
 
-int	gol_get_at_s(t_gol *gol, size_t x, size_t y)
+int	gol_get_at_s(t_gol *gol, ssize_t x, ssize_t y)
 {
-	if (x < 0 || x >= gol->width || y < 0 || y >= gol->height)
+	if (x < 0 || (size_t)x >= gol->width || y < 0 || (size_t)y >= gol->height)
 		return (0);
 	return (gol_get_at(gol, x, y));
 }
