+++ srcs/gol_set_at.c
@@ -6,7 +6,7 @@
 /*   By: gcros <gcros@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/22 21:30:48 by gcros             #+#    #+#             */
-/*   Updated: 2025/03/18 16:55:01 by gcros            ###   ########.fr       */
+/*   Updated: 2025/08/17 22:00:56 by gcros            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@ void	gol_set_at(t_gol *gol, size_t x, size_t y, int value)
 	gol->cells[y * gol->width + x].value = value;
 }
 
-void	gol_set_at_s(t_gol *gol, size_t x, size_t y, int value)
+void	gol_set_at_s(t_gol *gol, ssize_t x, ssize_t y, int value)
 {
-	if (x < 0 || x >= gol->width || y < 0 || y >= gol->height)
+	if (x < 0 || (size_t)x >= gol->width || y < 0 || (size_t)y >= gol->height)
 		return ;
 	return (gol_set_at(gol, x, y, value));
 }
